core/sphinx: Clean up tests, remove code duplication
